chore(cli): convert CLI test files from JS to TypeScript (batch 3)
Converts 20 test files in packages/cli/src/ from .js to .ts.
Includes proper TypeScript typing for vi.mock factories,
importOriginal<>() generics, and vi.mocked() for console spies.
