+++ packages/cli/src/commands/destroy/sdl/__tests__/sdl.test.ts
@@ -1,20 +1,21 @@
 globalThis.__dirname = __dirname
+
 import fs from 'node:fs'
 
 import { vol } from 'memfs'
-import { vi, describe, beforeEach, afterEach, test, expect } from 'vitest'
+import { vi, beforeEach, afterEach, test, expect, describe } from 'vitest'
 
 import '../../../../lib/test'
 
 import { getDefaultArgs } from '../../../../lib/index.js'
-import { builder } from '../../../generate/service/service.js'
-import { files } from '../../../generate/service/serviceHandler.js'
-import { tasks } from '../serviceHandler.js'
+import { builder } from '../../../generate/sdl/sdl.js'
+import { files } from '../../../generate/sdl/sdlHandler.js'
+import { tasks } from '../sdlHandler.js'
 
 vi.mock('node:fs')
 
 vi.mock('../../../../lib', async (importOriginal) => {
-  const originalLib = await importOriginal()
+  const originalLib = await importOriginal<typeof import('../../../../lib/index.js')>()
   return {
     ...originalLib,
     generateTemplate: () => '',
@@ -22,44 +23,37 @@ vi.mock('../../../../lib', async (importOriginal) => {
 })
 
 vi.mock('../../../../lib/schemaHelpers', async (importOriginal) => {
-  const originalSchemaHelpers = await importOriginal()
-  const path = require('path')
+  const originalSchemaHelpers = await importOriginal<typeof import('../../../../lib/schemaHelpers.js')>()
+  const { join } = await import('node:path')
+  const { readFileSync } = await vi.importActual<typeof import('node:fs')>('node:fs')
   return {
     ...originalSchemaHelpers,
     getSchema: () =>
-      require(path.join(globalThis.__dirname, 'fixtures', 'post.json')),
+      JSON.parse(
+        readFileSync(join(globalThis.__dirname, 'fixtures', 'post.json'), 'utf-8'),
+      ),
   }
 })
 
-describe('rw destroy service', () => {
-  beforeEach(() => {
-    vi.spyOn(console, 'info').mockImplementation(() => {})
-    vi.spyOn(console, 'log').mockImplementation(() => {})
-  })
-
+describe('rw destroy sdl', () => {
   afterEach(() => {
     vol.reset()
     vi.spyOn(fs, 'unlinkSync').mockClear()
-    console.info.mockRestore()
-    console.log.mockRestore()
   })
 
   describe('for javascript files', () => {
     beforeEach(async () => {
-      vol.fromJSON(await files({ ...getDefaultArgs(builder), name: 'User' }))
+      vol.fromJSON(await files({ ...getDefaultArgs(builder), name: 'Post' }))
     })
-    test('destroys service files', async () => {
+
+    test('destroys sdl files', async () => {
       const unlinkSpy = vi.spyOn(fs, 'unlinkSync')
-      const t = tasks({
-        componentName: 'service',
-        filesFn: files,
-        name: 'User',
-      })
+      const t = tasks({ model: 'Post' })
       t.options.renderer = 'silent'
 
-      return t.run().then(async () => {
+      return t.tasks[0].run().then(async () => {
         const generatedFiles = Object.keys(
-          await files({ ...getDefaultArgs(builder), name: 'User' }),
+          await files({ ...getDefaultArgs(builder), name: 'Post' }),
         )
         expect(generatedFiles.length).toEqual(unlinkSpy.mock.calls.length)
         generatedFiles.forEach((f) => expect(unlinkSpy).toHaveBeenCalledWith(f))
@@ -73,26 +67,22 @@ describe('rw destroy service', () => {
         await files({
           ...getDefaultArgs(builder),
           typescript: true,
-          name: 'User',
+          name: 'Post',
         }),
       )
     })
 
-    test('destroys service files', async () => {
+    test('destroys sdl files', async () => {
       const unlinkSpy = vi.spyOn(fs, 'unlinkSync')
-      const t = tasks({
-        componentName: 'service',
-        filesFn: files,
-        name: 'User',
-      })
+      const t = tasks({ model: 'Post' })
       t.options.renderer = 'silent'
 
-      return t.run().then(async () => {
+      return t.tasks[0].run().then(async () => {
         const generatedFiles = Object.keys(
           await files({
             ...getDefaultArgs(builder),
             typescript: true,
-            name: 'User',
+            name: 'Post',
           }),
         )
         expect(generatedFiles.length).toEqual(unlinkSpy.mock.calls.length)
