+++ packages/cli/src/commands/generate/__tests__/helpers.test.ts
