+++ packages/cli/src/commands/generate/scaffold/__tests__/scaffoldPathMultiword.test.ts
@@ -1,5 +1,5 @@
 globalThis.__dirname = __dirname
-import path from 'path'
+import path from 'node:path'
 
 import { vol, fs as memfs } from 'memfs'
 import { ufs } from 'unionfs'
@@ -12,7 +12,7 @@ import * as scaffoldHandler from '../scaffoldHandler.js'
 vi.mock('node:fs', async (importOriginal) => {
   const { wrapFsForUnionfs } =
     await import('../../../../__tests__/ufsFsProxy.js')
-  ufs.use(wrapFsForUnionfs(await importOriginal())).use(memfs)
+  ufs.use(wrapFsForUnionfs(await importOriginal<typeof import('node:fs')>())).use(memfs)
   return { ...ufs, default: { ...ufs } }
 })
 vi.mock('execa')
@@ -29,7 +29,7 @@ describe('AdminPages/Post', () => {
       model: 'Post',
       path: 'AdminPages',
       tests: true,
-      nestScaffoldByModel: false,
+      nestScaffoldByModel: true,
     })
   })
 
@@ -53,7 +53,7 @@ describe('AdminPages/Post', () => {
     test('creates a edit page', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
         ),
       ])
     })
@@ -62,18 +62,18 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
           )
         ],
       ).toMatch(
-        `import EditPostCell from 'src/components/AdminPages/EditPostCell'`,
+        `import EditPostCell from 'src/components/AdminPages/Post/EditPostCell'`,
       )
     })
 
     test('creates a index page', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
         ),
       ])
     })
@@ -82,16 +82,18 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
           )
         ],
-      ).toMatch(`import PostsCell from 'src/components/AdminPages/PostsCell'`)
+      ).toMatch(
+        `import PostsCell from 'src/components/AdminPages/Post/PostsCell'`,
+      )
     })
 
     test('creates a new page', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
         ),
       ])
     })
@@ -100,16 +102,16 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
           )
         ],
-      ).toMatch(`import NewPost from 'src/components/AdminPages/NewPost'`)
+      ).toMatch(`import NewPost from 'src/components/AdminPages/Post/NewPost'`)
     })
 
     test('creates a show page', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
         ),
       ])
     })
@@ -118,10 +120,12 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
           )
         ],
-      ).toMatch(`import PostCell from 'src/components/AdminPages/PostCell'`)
+      ).toMatch(
+        `import PostCell from 'src/components/AdminPages/Post/PostCell'`,
+      )
     })
 
     // Cells
@@ -129,7 +133,7 @@ describe('AdminPages/Post', () => {
     test('creates an edit cell', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
         ),
       ])
     })
@@ -138,16 +142,18 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
           )
         ],
-      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
+      ).toMatch(
+        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
+      )
     })
 
     test('creates an index cell', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
         ),
       ])
     })
@@ -156,16 +162,16 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
           )
         ],
-      ).toMatch(`import Posts from 'src/components/AdminPages/Posts'`)
+      ).toMatch(`import Posts from 'src/components/AdminPages/Post/Posts'`)
     })
 
     test('creates a show cell', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
         ),
       ])
     })
@@ -174,10 +180,10 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
           )
         ],
-      ).toMatch(`import Post from 'src/components/AdminPages/Post'`)
+      ).toMatch(`import Post from 'src/components/AdminPages/Post/Post'`)
     })
 
     // Components
@@ -185,7 +191,7 @@ describe('AdminPages/Post', () => {
     test('creates a form component', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/PostForm/PostForm.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/PostForm/PostForm.jsx',
         ),
       ])
     })
@@ -193,7 +199,7 @@ describe('AdminPages/Post', () => {
     test('creates an index component', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Posts/Posts.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/Posts/Posts.jsx',
         ),
       ])
     })
@@ -201,7 +207,7 @@ describe('AdminPages/Post', () => {
     test('creates a new component', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
         ),
       ])
     })
@@ -210,16 +216,18 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
           )
         ],
-      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
+      ).toMatch(
+        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
+      )
     })
 
     test('creates a show component', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/Post.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/Post/Post.jsx',
         ),
       ])
     })
@@ -233,13 +241,13 @@ describe('AdminPages/Post', () => {
         await scaffoldHandler.routes({
           model: 'Post',
           path: 'AdminPages',
-          nestScaffoldByModel: false,
+          nestScaffoldByModel: true,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/posts/new" page={AdminPagesNewPostPage} name="adminPagesNewPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesEditPostPage} name="adminPagesEditPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPage} name="adminPagesPost" />',
-        '<Route path="/admin-pages/posts" page={AdminPagesPostsPage} name="adminPagesPosts" />',
+        '<Route path="/admin-pages/posts/new" page={AdminPagesPostNewPostPage} name="adminPagesNewPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesPostEditPostPage} name="adminPagesEditPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPostPage} name="adminPagesPost" />',
+        '<Route path="/admin-pages/posts" page={AdminPagesPostPostsPage} name="adminPagesPosts" />',
       ])
     })
 
@@ -248,13 +256,13 @@ describe('AdminPages/Post', () => {
         await scaffoldHandler.routes({
           model: 'UserProfile',
           path: 'AdminPages',
-          nestScaffoldByModel: false,
+          nestScaffoldByModel: true,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesNewUserProfilePage} name="adminPagesNewUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesEditUserProfilePage} name="adminPagesEditUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfilePage} name="adminPagesUserProfile" />',
-        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfilesPage} name="adminPagesUserProfiles" />',
+        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesUserProfileNewUserProfilePage} name="adminPagesNewUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesUserProfileEditUserProfilePage} name="adminPagesEditUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfileUserProfilePage} name="adminPagesUserProfile" />',
+        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfileUserProfilesPage} name="adminPagesUserProfiles" />',
       ])
     })
   })
@@ -265,13 +273,13 @@ describe('AdminPages/Post', () => {
         model: 'UserProfile',
         path: 'AdminPages',
         tests: false,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfilesCell/UserProfilesCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfilesCell/UserProfilesCell.jsx',
           )
         ]
 
@@ -285,13 +293,13 @@ describe('AdminPages/Post', () => {
         model: 'UserProfile',
         path: 'AdminPages',
         tests: false,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfileCell/UserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfileCell/UserProfileCell.jsx',
           )
         ]
 
@@ -305,13 +313,13 @@ describe('AdminPages/Post', () => {
         model: 'UserProfile',
         path: 'AdminPages',
         tests: false,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
           )
         ]
 
@@ -327,12 +335,12 @@ describe('AdminPages/Post', () => {
         model: 'UserProfile',
         path: 'AdminPages',
         tests: false,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/NewUserProfile/NewUserProfile.jsx',
+          '/path/to/project/web/src/components/AdminPages/UserProfile/NewUserProfile/NewUserProfile.jsx',
         ),
       ])
     })
@@ -342,12 +350,12 @@ describe('AdminPages/Post', () => {
         model: 'UserProfile',
         path: 'AdminPages',
         tests: false,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
         ),
       ])
     })
@@ -362,7 +370,7 @@ describe('admin-pages/Post', () => {
       model: 'Post',
       path: 'admin-pages',
       tests: true,
-      nestScaffoldByModel: false,
+      nestScaffoldByModel: true,
     })
   })
 
@@ -386,7 +394,7 @@ describe('admin-pages/Post', () => {
     test('creates a edit page', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
         ),
       ])
     })
@@ -395,18 +403,18 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
           )
         ],
       ).toMatch(
-        `import EditPostCell from 'src/components/AdminPages/EditPostCell'`,
+        `import EditPostCell from 'src/components/AdminPages/Post/EditPostCell'`,
       )
     })
 
     test('creates a index page', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
         ),
       ])
     })
@@ -415,16 +423,18 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
           )
         ],
-      ).toMatch(`import PostsCell from 'src/components/AdminPages/PostsCell'`)
+      ).toMatch(
+        `import PostsCell from 'src/components/AdminPages/Post/PostsCell'`,
+      )
     })
 
     test('creates a new page', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
         ),
       ])
     })
@@ -433,16 +443,16 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
           )
         ],
-      ).toMatch(`import NewPost from 'src/components/AdminPages/NewPost'`)
+      ).toMatch(`import NewPost from 'src/components/AdminPages/Post/NewPost'`)
     })
 
     test('creates a show page', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
         ),
       ])
     })
@@ -451,10 +461,12 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
           )
         ],
-      ).toMatch(`import PostCell from 'src/components/AdminPages/PostCell'`)
+      ).toMatch(
+        `import PostCell from 'src/components/AdminPages/Post/PostCell'`,
+      )
     })
 
     // Cells
@@ -462,7 +474,7 @@ describe('admin-pages/Post', () => {
     test('creates an edit cell', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
         ),
       ])
     })
@@ -471,16 +483,18 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
           )
         ],
-      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
+      ).toMatch(
+        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
+      )
     })
 
     test('creates an index cell', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
         ),
       ])
     })
@@ -489,16 +503,16 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
           )
         ],
-      ).toMatch(`import Posts from 'src/components/AdminPages/Posts'`)
+      ).toMatch(`import Posts from 'src/components/AdminPages/Post/Posts'`)
     })
 
     test('creates a show cell', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
         ),
       ])
     })
@@ -507,10 +521,10 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
           )
         ],
-      ).toMatch(`import Post from 'src/components/AdminPages/Post'`)
+      ).toMatch(`import Post from 'src/components/AdminPages/Post/Post'`)
     })
 
     // Components
@@ -518,7 +532,7 @@ describe('admin-pages/Post', () => {
     test('creates a form component', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/PostForm/PostForm.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/PostForm/PostForm.jsx',
         ),
       ])
     })
@@ -526,7 +540,7 @@ describe('admin-pages/Post', () => {
     test('creates an index component', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Posts/Posts.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/Posts/Posts.jsx',
         ),
       ])
     })
@@ -534,7 +548,7 @@ describe('admin-pages/Post', () => {
     test('creates a new component', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
         ),
       ])
     })
@@ -543,16 +557,18 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
           )
         ],
-      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
+      ).toMatch(
+        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
+      )
     })
 
     test('creates a show component', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/Post.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/Post/Post.jsx',
         ),
       ])
     })
@@ -564,30 +580,30 @@ describe('admin-pages/Post', () => {
     test('creates a single-word name routes', async () => {
       expect(
         await scaffoldHandler.routes({
-          model: 'Post',
+          model: 'post',
           path: 'admin-pages',
-          nestScaffoldByModel: false,
+          nestScaffoldByModel: true,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/posts/new" page={AdminPagesNewPostPage} name="adminPagesNewPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesEditPostPage} name="adminPagesEditPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPage} name="adminPagesPost" />',
-        '<Route path="/admin-pages/posts" page={AdminPagesPostsPage} name="adminPagesPosts" />',
+        '<Route path="/admin-pages/posts/new" page={AdminPagesPostNewPostPage} name="adminPagesNewPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesPostEditPostPage} name="adminPagesEditPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPostPage} name="adminPagesPost" />',
+        '<Route path="/admin-pages/posts" page={AdminPagesPostPostsPage} name="adminPagesPosts" />',
       ])
     })
 
     test('creates a multi-word name routes', async () => {
       expect(
         await scaffoldHandler.routes({
-          model: 'UserProfile',
+          model: 'userProfile',
           path: 'admin-pages',
-          nestScaffoldByModel: false,
+          nestScaffoldByModel: true,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesNewUserProfilePage} name="adminPagesNewUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesEditUserProfilePage} name="adminPagesEditUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfilePage} name="adminPagesUserProfile" />',
-        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfilesPage} name="adminPagesUserProfiles" />',
+        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesUserProfileNewUserProfilePage} name="adminPagesNewUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesUserProfileEditUserProfilePage} name="adminPagesEditUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfileUserProfilePage} name="adminPagesUserProfile" />',
+        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfileUserProfilesPage} name="adminPagesUserProfiles" />',
       ])
     })
   })
@@ -598,13 +614,13 @@ describe('admin-pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: false,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfilesCell/UserProfilesCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfilesCell/UserProfilesCell.jsx',
           )
         ]
 
@@ -618,13 +634,13 @@ describe('admin-pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: false,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfileCell/UserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfileCell/UserProfileCell.jsx',
           )
         ]
 
@@ -638,13 +654,13 @@ describe('admin-pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: false,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
           )
         ]
 
@@ -660,12 +676,12 @@ describe('admin-pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: false,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/NewUserProfile/NewUserProfile.jsx',
+          '/path/to/project/web/src/components/AdminPages/UserProfile/NewUserProfile/NewUserProfile.jsx',
         ),
       ])
     })
@@ -675,12 +691,12 @@ describe('admin-pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: false,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
         ),
       ])
     })
@@ -695,7 +711,7 @@ describe('admin_pages/Post', () => {
       model: 'Post',
       path: 'admin_pages',
       tests: true,
-      nestScaffoldByModel: false,
+      nestScaffoldByModel: true,
     })
   })
 
@@ -719,7 +735,7 @@ describe('admin_pages/Post', () => {
     test('creates a edit page', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
         ),
       ])
     })
@@ -728,18 +744,18 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
           )
         ],
       ).toMatch(
-        `import EditPostCell from 'src/components/AdminPages/EditPostCell'`,
+        `import EditPostCell from 'src/components/AdminPages/Post/EditPostCell'`,
       )
     })
 
     test('creates a index page', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
         ),
       ])
     })
@@ -748,16 +764,18 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
           )
         ],
-      ).toMatch(`import PostsCell from 'src/components/AdminPages/PostsCell'`)
+      ).toMatch(
+        `import PostsCell from 'src/components/AdminPages/Post/PostsCell'`,
+      )
     })
 
     test('creates a new page', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
         ),
       ])
     })
@@ -766,16 +784,16 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
           )
         ],
-      ).toMatch(`import NewPost from 'src/components/AdminPages/NewPost'`)
+      ).toMatch(`import NewPost from 'src/components/AdminPages/Post/NewPost'`)
     })
 
     test('creates a show page', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
         ),
       ])
     })
@@ -784,10 +802,12 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
           )
         ],
-      ).toMatch(`import PostCell from 'src/components/AdminPages/PostCell'`)
+      ).toMatch(
+        `import PostCell from 'src/components/AdminPages/Post/PostCell'`,
+      )
     })
 
     // Cells
@@ -795,7 +815,7 @@ describe('admin_pages/Post', () => {
     test('creates an edit cell', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
         ),
       ])
     })
@@ -804,16 +824,18 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
           )
         ],
-      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
+      ).toMatch(
+        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
+      )
     })
 
     test('creates an index cell', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
         ),
       ])
     })
@@ -822,16 +844,16 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
           )
         ],
-      ).toMatch(`import Posts from 'src/components/AdminPages/Posts'`)
+      ).toMatch(`import Posts from 'src/components/AdminPages/Post/Posts'`)
     })
 
     test('creates a show cell', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
         ),
       ])
     })
@@ -840,10 +862,10 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
           )
         ],
-      ).toMatch(`import Post from 'src/components/AdminPages/Post'`)
+      ).toMatch(`import Post from 'src/components/AdminPages/Post/Post'`)
     })
 
     // Components
@@ -851,7 +873,7 @@ describe('admin_pages/Post', () => {
     test('creates a form component', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/PostForm/PostForm.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/PostForm/PostForm.jsx',
         ),
       ])
     })
@@ -859,7 +881,7 @@ describe('admin_pages/Post', () => {
     test('creates an index component', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Posts/Posts.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/Posts/Posts.jsx',
         ),
       ])
     })
@@ -867,7 +889,7 @@ describe('admin_pages/Post', () => {
     test('creates a new component', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
         ),
       ])
     })
@@ -876,16 +898,18 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
+            '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
           )
         ],
-      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
+      ).toMatch(
+        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
+      )
     })
 
     test('creates a show component', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/Post.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/Post/Post.jsx',
         ),
       ])
     })
@@ -899,13 +923,13 @@ describe('admin_pages/Post', () => {
         await scaffoldHandler.routes({
           model: 'Post',
           path: 'admin_pages',
-          nestScaffoldByModel: false,
+          nestScaffoldByModel: true,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/posts/new" page={AdminPagesNewPostPage} name="adminPagesNewPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesEditPostPage} name="adminPagesEditPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPage} name="adminPagesPost" />',
-        '<Route path="/admin-pages/posts" page={AdminPagesPostsPage} name="adminPagesPosts" />',
+        '<Route path="/admin-pages/posts/new" page={AdminPagesPostNewPostPage} name="adminPagesNewPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesPostEditPostPage} name="adminPagesEditPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPostPage} name="adminPagesPost" />',
+        '<Route path="/admin-pages/posts" page={AdminPagesPostPostsPage} name="adminPagesPosts" />',
       ])
     })
 
@@ -914,13 +938,13 @@ describe('admin_pages/Post', () => {
         await scaffoldHandler.routes({
           model: 'UserProfile',
           path: 'admin_pages',
-          nestScaffoldByModel: false,
+          nestScaffoldByModel: true,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesNewUserProfilePage} name="adminPagesNewUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesEditUserProfilePage} name="adminPagesEditUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfilePage} name="adminPagesUserProfile" />',
-        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfilesPage} name="adminPagesUserProfiles" />',
+        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesUserProfileNewUserProfilePage} name="adminPagesNewUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesUserProfileEditUserProfilePage} name="adminPagesEditUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfileUserProfilePage} name="adminPagesUserProfile" />',
+        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfileUserProfilesPage} name="adminPagesUserProfiles" />',
       ])
     })
   })
@@ -931,13 +955,13 @@ describe('admin_pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: true,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfilesCell/UserProfilesCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfilesCell/UserProfilesCell.jsx',
           )
         ]
 
@@ -951,13 +975,13 @@ describe('admin_pages/Post', () => {
         model: 'UserProfile',
         path: 'admin_pages',
         tests: true,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfileCell/UserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfileCell/UserProfileCell.jsx',
           )
         ]
 
@@ -971,13 +995,13 @@ describe('admin_pages/Post', () => {
         model: 'UserProfile',
         path: 'admin_pages',
         tests: true,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
           )
         ]
 
@@ -993,12 +1017,12 @@ describe('admin_pages/Post', () => {
         model: 'UserProfile',
         path: 'admin_pages',
         tests: true,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/NewUserProfile/NewUserProfile.jsx',
+          '/path/to/project/web/src/components/AdminPages/UserProfile/NewUserProfile/NewUserProfile.jsx',
         ),
       ])
     })
@@ -1008,12 +1032,12 @@ describe('admin_pages/Post', () => {
         model: 'UserProfile',
         path: 'admin_pages',
         tests: true,
-        nestScaffoldByModel: false,
+        nestScaffoldByModel: true,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
         ),
       ])
     })
