+++ packages/cli/src/commands/generate/scaffold/__tests__/scaffoldPathMultiwordNoNest.test.ts
@@ -12,7 +12,7 @@ import * as scaffoldHandler from '../scaffoldHandler.js'
 vi.mock('node:fs', async (importOriginal) => {
   const { wrapFsForUnionfs } =
     await import('../../../../__tests__/ufsFsProxy.js')
-  ufs.use(wrapFsForUnionfs(await importOriginal())).use(memfs)
+  ufs.use(wrapFsForUnionfs(await importOriginal<typeof import('node:fs')>())).use(memfs)
   return { ...ufs, default: { ...ufs } }
 })
 vi.mock('execa')
@@ -29,7 +29,7 @@ describe('AdminPages/Post', () => {
       model: 'Post',
       path: 'AdminPages',
       tests: true,
-      nestScaffoldByModel: true,
+      nestScaffoldByModel: false,
     })
   })
 
@@ -53,7 +53,7 @@ describe('AdminPages/Post', () => {
     test('creates a edit page', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
         ),
       ])
     })
@@ -62,18 +62,18 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
           )
         ],
       ).toMatch(
-        `import EditPostCell from 'src/components/AdminPages/Post/EditPostCell'`,
+        `import EditPostCell from 'src/components/AdminPages/EditPostCell'`,
       )
     })
 
     test('creates a index page', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
         ),
       ])
     })
@@ -82,18 +82,16 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
           )
         ],
-      ).toMatch(
-        `import PostsCell from 'src/components/AdminPages/Post/PostsCell'`,
-      )
+      ).toMatch(`import PostsCell from 'src/components/AdminPages/PostsCell'`)
     })
 
     test('creates a new page', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
         ),
       ])
     })
@@ -102,16 +100,16 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
           )
         ],
-      ).toMatch(`import NewPost from 'src/components/AdminPages/Post/NewPost'`)
+      ).toMatch(`import NewPost from 'src/components/AdminPages/NewPost'`)
     })
 
     test('creates a show page', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
         ),
       ])
     })
@@ -120,12 +118,10 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
           )
         ],
-      ).toMatch(
-        `import PostCell from 'src/components/AdminPages/Post/PostCell'`,
-      )
+      ).toMatch(`import PostCell from 'src/components/AdminPages/PostCell'`)
     })
 
     // Cells
@@ -133,7 +129,7 @@ describe('AdminPages/Post', () => {
     test('creates an edit cell', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
         ),
       ])
     })
@@ -142,18 +138,16 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
           )
         ],
-      ).toMatch(
-        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
-      )
+      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
     })
 
     test('creates an index cell', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
         ),
       ])
     })
@@ -162,16 +156,16 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
           )
         ],
-      ).toMatch(`import Posts from 'src/components/AdminPages/Post/Posts'`)
+      ).toMatch(`import Posts from 'src/components/AdminPages/Posts'`)
     })
 
     test('creates a show cell', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
         ),
       ])
     })
@@ -180,10 +174,10 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
           )
         ],
-      ).toMatch(`import Post from 'src/components/AdminPages/Post/Post'`)
+      ).toMatch(`import Post from 'src/components/AdminPages/Post'`)
     })
 
     // Components
@@ -191,7 +185,7 @@ describe('AdminPages/Post', () => {
     test('creates a form component', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/PostForm/PostForm.jsx',
+          '/path/to/project/web/src/components/AdminPages/PostForm/PostForm.jsx',
         ),
       ])
     })
@@ -199,7 +193,7 @@ describe('AdminPages/Post', () => {
     test('creates an index component', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/Posts/Posts.jsx',
+          '/path/to/project/web/src/components/AdminPages/Posts/Posts.jsx',
         ),
       ])
     })
@@ -207,7 +201,7 @@ describe('AdminPages/Post', () => {
     test('creates a new component', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
+          '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
         ),
       ])
     })
@@ -216,18 +210,16 @@ describe('AdminPages/Post', () => {
       expect(
         filesMultiwordUpper[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
+            '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
           )
         ],
-      ).toMatch(
-        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
-      )
+      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
     })
 
     test('creates a show component', async () => {
       expect(filesMultiwordUpper).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/Post/Post.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/Post.jsx',
         ),
       ])
     })
@@ -241,13 +233,13 @@ describe('AdminPages/Post', () => {
         await scaffoldHandler.routes({
           model: 'Post',
           path: 'AdminPages',
-          nestScaffoldByModel: true,
+          nestScaffoldByModel: false,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/posts/new" page={AdminPagesPostNewPostPage} name="adminPagesNewPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesPostEditPostPage} name="adminPagesEditPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPostPage} name="adminPagesPost" />',
-        '<Route path="/admin-pages/posts" page={AdminPagesPostPostsPage} name="adminPagesPosts" />',
+        '<Route path="/admin-pages/posts/new" page={AdminPagesNewPostPage} name="adminPagesNewPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesEditPostPage} name="adminPagesEditPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPage} name="adminPagesPost" />',
+        '<Route path="/admin-pages/posts" page={AdminPagesPostsPage} name="adminPagesPosts" />',
       ])
     })
 
@@ -256,13 +248,13 @@ describe('AdminPages/Post', () => {
         await scaffoldHandler.routes({
           model: 'UserProfile',
           path: 'AdminPages',
-          nestScaffoldByModel: true,
+          nestScaffoldByModel: false,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesUserProfileNewUserProfilePage} name="adminPagesNewUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesUserProfileEditUserProfilePage} name="adminPagesEditUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfileUserProfilePage} name="adminPagesUserProfile" />',
-        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfileUserProfilesPage} name="adminPagesUserProfiles" />',
+        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesNewUserProfilePage} name="adminPagesNewUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesEditUserProfilePage} name="adminPagesEditUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfilePage} name="adminPagesUserProfile" />',
+        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfilesPage} name="adminPagesUserProfiles" />',
       ])
     })
   })
@@ -273,13 +265,13 @@ describe('AdminPages/Post', () => {
         model: 'UserProfile',
         path: 'AdminPages',
         tests: false,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfilesCell/UserProfilesCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfilesCell/UserProfilesCell.jsx',
           )
         ]
 
@@ -293,13 +285,13 @@ describe('AdminPages/Post', () => {
         model: 'UserProfile',
         path: 'AdminPages',
         tests: false,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfileCell/UserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfileCell/UserProfileCell.jsx',
           )
         ]
 
@@ -313,13 +305,13 @@ describe('AdminPages/Post', () => {
         model: 'UserProfile',
         path: 'AdminPages',
         tests: false,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
           )
         ]
 
@@ -335,12 +327,12 @@ describe('AdminPages/Post', () => {
         model: 'UserProfile',
         path: 'AdminPages',
         tests: false,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/UserProfile/NewUserProfile/NewUserProfile.jsx',
+          '/path/to/project/web/src/components/AdminPages/NewUserProfile/NewUserProfile.jsx',
         ),
       ])
     })
@@ -350,12 +342,12 @@ describe('AdminPages/Post', () => {
         model: 'UserProfile',
         path: 'AdminPages',
         tests: false,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
         ),
       ])
     })
@@ -370,7 +362,7 @@ describe('admin-pages/Post', () => {
       model: 'Post',
       path: 'admin-pages',
       tests: true,
-      nestScaffoldByModel: true,
+      nestScaffoldByModel: false,
     })
   })
 
@@ -394,7 +386,7 @@ describe('admin-pages/Post', () => {
     test('creates a edit page', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
         ),
       ])
     })
@@ -403,18 +395,18 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
           )
         ],
       ).toMatch(
-        `import EditPostCell from 'src/components/AdminPages/Post/EditPostCell'`,
+        `import EditPostCell from 'src/components/AdminPages/EditPostCell'`,
       )
     })
 
     test('creates a index page', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
         ),
       ])
     })
@@ -423,18 +415,16 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
           )
         ],
-      ).toMatch(
-        `import PostsCell from 'src/components/AdminPages/Post/PostsCell'`,
-      )
+      ).toMatch(`import PostsCell from 'src/components/AdminPages/PostsCell'`)
     })
 
     test('creates a new page', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
         ),
       ])
     })
@@ -443,16 +433,16 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
           )
         ],
-      ).toMatch(`import NewPost from 'src/components/AdminPages/Post/NewPost'`)
+      ).toMatch(`import NewPost from 'src/components/AdminPages/NewPost'`)
     })
 
     test('creates a show page', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
         ),
       ])
     })
@@ -461,12 +451,10 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
           )
         ],
-      ).toMatch(
-        `import PostCell from 'src/components/AdminPages/Post/PostCell'`,
-      )
+      ).toMatch(`import PostCell from 'src/components/AdminPages/PostCell'`)
     })
 
     // Cells
@@ -474,7 +462,7 @@ describe('admin-pages/Post', () => {
     test('creates an edit cell', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
         ),
       ])
     })
@@ -483,18 +471,16 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
           )
         ],
-      ).toMatch(
-        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
-      )
+      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
     })
 
     test('creates an index cell', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
         ),
       ])
     })
@@ -503,16 +489,16 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
           )
         ],
-      ).toMatch(`import Posts from 'src/components/AdminPages/Post/Posts'`)
+      ).toMatch(`import Posts from 'src/components/AdminPages/Posts'`)
     })
 
     test('creates a show cell', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
         ),
       ])
     })
@@ -521,10 +507,10 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
           )
         ],
-      ).toMatch(`import Post from 'src/components/AdminPages/Post/Post'`)
+      ).toMatch(`import Post from 'src/components/AdminPages/Post'`)
     })
 
     // Components
@@ -532,7 +518,7 @@ describe('admin-pages/Post', () => {
     test('creates a form component', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/PostForm/PostForm.jsx',
+          '/path/to/project/web/src/components/AdminPages/PostForm/PostForm.jsx',
         ),
       ])
     })
@@ -540,7 +526,7 @@ describe('admin-pages/Post', () => {
     test('creates an index component', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/Posts/Posts.jsx',
+          '/path/to/project/web/src/components/AdminPages/Posts/Posts.jsx',
         ),
       ])
     })
@@ -548,7 +534,7 @@ describe('admin-pages/Post', () => {
     test('creates a new component', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
+          '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
         ),
       ])
     })
@@ -557,18 +543,16 @@ describe('admin-pages/Post', () => {
       expect(
         filesMultiwordDash[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
+            '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
           )
         ],
-      ).toMatch(
-        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
-      )
+      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
     })
 
     test('creates a show component', async () => {
       expect(filesMultiwordDash).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/Post/Post.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/Post.jsx',
         ),
       ])
     })
@@ -580,30 +564,30 @@ describe('admin-pages/Post', () => {
     test('creates a single-word name routes', async () => {
       expect(
         await scaffoldHandler.routes({
-          model: 'post',
+          model: 'Post',
           path: 'admin-pages',
-          nestScaffoldByModel: true,
+          nestScaffoldByModel: false,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/posts/new" page={AdminPagesPostNewPostPage} name="adminPagesNewPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesPostEditPostPage} name="adminPagesEditPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPostPage} name="adminPagesPost" />',
-        '<Route path="/admin-pages/posts" page={AdminPagesPostPostsPage} name="adminPagesPosts" />',
+        '<Route path="/admin-pages/posts/new" page={AdminPagesNewPostPage} name="adminPagesNewPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesEditPostPage} name="adminPagesEditPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPage} name="adminPagesPost" />',
+        '<Route path="/admin-pages/posts" page={AdminPagesPostsPage} name="adminPagesPosts" />',
       ])
     })
 
     test('creates a multi-word name routes', async () => {
       expect(
         await scaffoldHandler.routes({
-          model: 'userProfile',
+          model: 'UserProfile',
           path: 'admin-pages',
-          nestScaffoldByModel: true,
+          nestScaffoldByModel: false,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesUserProfileNewUserProfilePage} name="adminPagesNewUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesUserProfileEditUserProfilePage} name="adminPagesEditUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfileUserProfilePage} name="adminPagesUserProfile" />',
-        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfileUserProfilesPage} name="adminPagesUserProfiles" />',
+        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesNewUserProfilePage} name="adminPagesNewUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesEditUserProfilePage} name="adminPagesEditUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfilePage} name="adminPagesUserProfile" />',
+        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfilesPage} name="adminPagesUserProfiles" />',
       ])
     })
   })
@@ -614,13 +598,13 @@ describe('admin-pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: false,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfilesCell/UserProfilesCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfilesCell/UserProfilesCell.jsx',
           )
         ]
 
@@ -634,13 +618,13 @@ describe('admin-pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: false,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfileCell/UserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfileCell/UserProfileCell.jsx',
           )
         ]
 
@@ -654,13 +638,13 @@ describe('admin-pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: false,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
           )
         ]
 
@@ -676,12 +660,12 @@ describe('admin-pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: false,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/UserProfile/NewUserProfile/NewUserProfile.jsx',
+          '/path/to/project/web/src/components/AdminPages/NewUserProfile/NewUserProfile.jsx',
         ),
       ])
     })
@@ -691,12 +675,12 @@ describe('admin-pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: false,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
         ),
       ])
     })
@@ -711,7 +695,7 @@ describe('admin_pages/Post', () => {
       model: 'Post',
       path: 'admin_pages',
       tests: true,
-      nestScaffoldByModel: true,
+      nestScaffoldByModel: false,
     })
   })
 
@@ -735,7 +719,7 @@ describe('admin_pages/Post', () => {
     test('creates a edit page', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
         ),
       ])
     })
@@ -744,18 +728,18 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/EditPostPage/EditPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/EditPostPage/EditPostPage.jsx',
           )
         ],
       ).toMatch(
-        `import EditPostCell from 'src/components/AdminPages/Post/EditPostCell'`,
+        `import EditPostCell from 'src/components/AdminPages/EditPostCell'`,
       )
     })
 
     test('creates a index page', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
         ),
       ])
     })
@@ -764,18 +748,16 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/PostsPage/PostsPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/PostsPage/PostsPage.jsx',
           )
         ],
-      ).toMatch(
-        `import PostsCell from 'src/components/AdminPages/Post/PostsCell'`,
-      )
+      ).toMatch(`import PostsCell from 'src/components/AdminPages/PostsCell'`)
     })
 
     test('creates a new page', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
         ),
       ])
     })
@@ -784,16 +766,16 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/NewPostPage/NewPostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/NewPostPage/NewPostPage.jsx',
           )
         ],
-      ).toMatch(`import NewPost from 'src/components/AdminPages/Post/NewPost'`)
+      ).toMatch(`import NewPost from 'src/components/AdminPages/NewPost'`)
     })
 
     test('creates a show page', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
+          '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
         ),
       ])
     })
@@ -802,12 +784,10 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/pages/AdminPages/Post/PostPage/PostPage.jsx',
+            '/path/to/project/web/src/pages/AdminPages/PostPage/PostPage.jsx',
           )
         ],
-      ).toMatch(
-        `import PostCell from 'src/components/AdminPages/Post/PostCell'`,
-      )
+      ).toMatch(`import PostCell from 'src/components/AdminPages/PostCell'`)
     })
 
     // Cells
@@ -815,7 +795,7 @@ describe('admin_pages/Post', () => {
     test('creates an edit cell', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
         ),
       ])
     })
@@ -824,18 +804,16 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/EditPostCell/EditPostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/EditPostCell/EditPostCell.jsx',
           )
         ],
-      ).toMatch(
-        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
-      )
+      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
     })
 
     test('creates an index cell', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
         ),
       ])
     })
@@ -844,16 +822,16 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/PostsCell/PostsCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/PostsCell/PostsCell.jsx',
           )
         ],
-      ).toMatch(`import Posts from 'src/components/AdminPages/Post/Posts'`)
+      ).toMatch(`import Posts from 'src/components/AdminPages/Posts'`)
     })
 
     test('creates a show cell', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
         ),
       ])
     })
@@ -862,10 +840,10 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/PostCell/PostCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/PostCell/PostCell.jsx',
           )
         ],
-      ).toMatch(`import Post from 'src/components/AdminPages/Post/Post'`)
+      ).toMatch(`import Post from 'src/components/AdminPages/Post'`)
     })
 
     // Components
@@ -873,7 +851,7 @@ describe('admin_pages/Post', () => {
     test('creates a form component', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/PostForm/PostForm.jsx',
+          '/path/to/project/web/src/components/AdminPages/PostForm/PostForm.jsx',
         ),
       ])
     })
@@ -881,7 +859,7 @@ describe('admin_pages/Post', () => {
     test('creates an index component', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/Posts/Posts.jsx',
+          '/path/to/project/web/src/components/AdminPages/Posts/Posts.jsx',
         ),
       ])
     })
@@ -889,7 +867,7 @@ describe('admin_pages/Post', () => {
     test('creates a new component', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
+          '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
         ),
       ])
     })
@@ -898,18 +876,16 @@ describe('admin_pages/Post', () => {
       expect(
         filesMultiwordUnderscore[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/Post/NewPost/NewPost.jsx',
+            '/path/to/project/web/src/components/AdminPages/NewPost/NewPost.jsx',
           )
         ],
-      ).toMatch(
-        `import PostForm from 'src/components/AdminPages/Post/PostForm'`,
-      )
+      ).toMatch(`import PostForm from 'src/components/AdminPages/PostForm'`)
     })
 
     test('creates a show component', async () => {
       expect(filesMultiwordUnderscore).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/Post/Post/Post.jsx',
+          '/path/to/project/web/src/components/AdminPages/Post/Post.jsx',
         ),
       ])
     })
@@ -923,13 +899,13 @@ describe('admin_pages/Post', () => {
         await scaffoldHandler.routes({
           model: 'Post',
           path: 'admin_pages',
-          nestScaffoldByModel: true,
+          nestScaffoldByModel: false,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/posts/new" page={AdminPagesPostNewPostPage} name="adminPagesNewPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesPostEditPostPage} name="adminPagesEditPost" />',
-        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPostPage} name="adminPagesPost" />',
-        '<Route path="/admin-pages/posts" page={AdminPagesPostPostsPage} name="adminPagesPosts" />',
+        '<Route path="/admin-pages/posts/new" page={AdminPagesNewPostPage} name="adminPagesNewPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}/edit" page={AdminPagesEditPostPage} name="adminPagesEditPost" />',
+        '<Route path="/admin-pages/posts/{id:Int}" page={AdminPagesPostPage} name="adminPagesPost" />',
+        '<Route path="/admin-pages/posts" page={AdminPagesPostsPage} name="adminPagesPosts" />',
       ])
     })
 
@@ -938,13 +914,13 @@ describe('admin_pages/Post', () => {
         await scaffoldHandler.routes({
           model: 'UserProfile',
           path: 'admin_pages',
-          nestScaffoldByModel: true,
+          nestScaffoldByModel: false,
         }),
       ).toEqual([
-        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesUserProfileNewUserProfilePage} name="adminPagesNewUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesUserProfileEditUserProfilePage} name="adminPagesEditUserProfile" />',
-        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfileUserProfilePage} name="adminPagesUserProfile" />',
-        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfileUserProfilesPage} name="adminPagesUserProfiles" />',
+        '<Route path="/admin-pages/user-profiles/new" page={AdminPagesNewUserProfilePage} name="adminPagesNewUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}/edit" page={AdminPagesEditUserProfilePage} name="adminPagesEditUserProfile" />',
+        '<Route path="/admin-pages/user-profiles/{id:Int}" page={AdminPagesUserProfilePage} name="adminPagesUserProfile" />',
+        '<Route path="/admin-pages/user-profiles" page={AdminPagesUserProfilesPage} name="adminPagesUserProfiles" />',
       ])
     })
   })
@@ -955,13 +931,13 @@ describe('admin_pages/Post', () => {
         model: 'UserProfile',
         path: 'admin-pages',
         tests: true,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfilesCell/UserProfilesCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfilesCell/UserProfilesCell.jsx',
           )
         ]
 
@@ -975,13 +951,13 @@ describe('admin_pages/Post', () => {
         model: 'UserProfile',
         path: 'admin_pages',
         tests: true,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfile/UserProfileCell/UserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/UserProfileCell/UserProfileCell.jsx',
           )
         ]
 
@@ -995,13 +971,13 @@ describe('admin_pages/Post', () => {
         model: 'UserProfile',
         path: 'admin_pages',
         tests: true,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       const cell =
         userProfileFiles[
           path.normalize(
-            '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
+            '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
           )
         ]
 
@@ -1017,12 +993,12 @@ describe('admin_pages/Post', () => {
         model: 'UserProfile',
         path: 'admin_pages',
         tests: true,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/UserProfile/NewUserProfile/NewUserProfile.jsx',
+          '/path/to/project/web/src/components/AdminPages/NewUserProfile/NewUserProfile.jsx',
         ),
       ])
     })
@@ -1032,12 +1008,12 @@ describe('admin_pages/Post', () => {
         model: 'UserProfile',
         path: 'admin_pages',
         tests: true,
-        nestScaffoldByModel: true,
+        nestScaffoldByModel: false,
       })
 
       expect(foreignKeyFiles).toHaveProperty([
         path.normalize(
-          '/path/to/project/web/src/components/AdminPages/UserProfile/EditUserProfileCell/EditUserProfileCell.jsx',
+          '/path/to/project/web/src/components/AdminPages/EditUserProfileCell/EditUserProfileCell.jsx',
         ),
       ])
     })
