+++ packages/cli/src/commands/generate/scaffold/__tests__/shouldUseTailwindCSS.test.ts
