+++ packages/cli/src/commands/generate/sdl/__tests__/sdl.test.ts
@@ -13,7 +13,7 @@ import '../../../../lib/test'
 vi.mock('node:fs', async (importOriginal) => {
   const { wrapFsForUnionfs } =
     await import('../../../../__tests__/ufsFsProxy.js')
-  ufs.use(wrapFsForUnionfs(await importOriginal())).use(memfs)
+  ufs.use(wrapFsForUnionfs(await importOriginal<typeof import('node:fs')>())).use(memfs)
 
   return { ...ufs, default: ufs }
 })
