+++ packages/cli/src/lib/__tests__/mergeConfig.test.ts
@@ -1,5 +1,5 @@
 import fs from 'node:fs'
-import path from 'path'
+import path from 'node:path'
 
 import { expect, it, describe, test } from 'vitest'
